refactor: change routes and imports
