+++ package.json
@@ -43,6 +43,7 @@
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "dotenv": "^16.0.1",
+    "express-winston": "^4.2.0",
     "http-status-codes": "^2.2.0",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
