+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { JwtModule } from '@nestjs/jwt';
-import { JwtMiddlewareModule } from './jwt.middleware.module';
+import { JwtMiddlewareModule } from './middleware/jwtMiddleware/jwt.middleware.module';
 import { AppController } from './app.controller';
-import { DeleteFavoritesMiddlewareModule } from './deleteFavoritesMiddleware.module';
+import { DeleteFavoritesMiddlewareModule } from './middleware/deleteMiddleware/deleteFavoritesMiddleware.module';
 import { FavoritesModule } from './favorites/favorites.module';
 import { AlbumModule } from './album/album.module';
 import { ArtistModule } from './artist/artist.module';
