+++ src/middleware/deleteMiddleware/deleteFavorites.middleware.spec.ts
