+++ src/middleware/deleteMiddleware/deleteFavoritesMiddleware.module.ts
@@ -1,4 +1,4 @@
-import { FavoritesModule } from './favorites/favorites.module';
+import { FavoritesModule } from '../../favorites/favorites.module';
 import { DeleteFavoritesMiddleware } from './deleteFavorites.middleware';
 import { MiddlewareConsumer, Module, RequestMethod } from '@nestjs/common';
